Extract classes from ToolStripManager
